+++ ydb/tests/stress/kv_volume/execution_context.cpp
@@ -1,0 +1,40 @@
+#include "execution_context.h"
+
+namespace NKvVolumeStress {
+
+TExecutionContext::TExecutionContext(
+    ui64 executionId,
+    TString actionName,
+    TPtr parent,
+    TKeyBucket* workerStorage)
+    : ExecutionId(executionId)
+    , ActionName(std::move(actionName))
+    , Parent(std::move(parent))
+    , WorkerStorage_(workerStorage)
+{
+}
+
+TExecutionContext::~TExecutionContext() {
+    if (!WorkerStorage_) {
+        return;
+    }
+
+    const TVector<std::pair<TString, TKeyInfo>> remaining = Keys_.Drain();
+    if (!remaining.empty()) {
+        WorkerStorage_->AddKeys(remaining);
+    }
+}
+
+void TExecutionContext::AddKey(const TString& key, const TKeyInfo& keyInfo) {
+    Keys_.AddKey(key, keyInfo);
+}
+
+void TExecutionContext::AddKeys(const TVector<std::pair<TString, TKeyInfo>>& keys) {
+    Keys_.AddKeys(keys);
+}
+
+TVector<std::pair<TString, TKeyInfo>> TExecutionContext::PickKeys(ui32 count, bool erase) {
+    return Keys_.PickKeys(count, erase);
+}
+
+} // namespace NKvVolumeStress
